+++ index.html
@@ -234,8 +234,7 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="js/functions.js"></script>
-  <script src="./js/main.js" defer></script>
+  <script src="js/main.js" type="module" defer></script>
 </body>
 
 </html>
